Implemented Interface for Binary Tree and Updated NamingConvention.md and TODO.md
In NamingConvention.md
=======================

In this I added the Interface convention

For Interface of Binary Tree
==============================

I used a little bit confusing idea, the base idea the data stored in the node may not alway point a
primitive data (by means itself). If it points to pointer well and good, but it is is stored for
a pathway for complex data (anything comprehendable) , that's why I added two function to the Interface
for it's node also include normal data.

Usually node :-

template<type>
node{ // Primitive node
    type data;
    ptr p,left,right;

//  Usually the programming interface provide these so no need to implement it by any mean
    type read();
    bool write(type);
}

Interface for For the node :-

template<gateway_type , actual_data_type>
node{ // complex node
    gateway_type data;
    ptr p,left,right;

//  These have to Implemented by some defined means as the computer can interpret these unknown conversion.
//  also the actual_data_type must be comprehenable by the system, and pre Implemented (user or otherwise)
    actual_data_type read();
    bool write(actual_data_type);
}

The complex node can be used as Primitive node by making the gateway_type and actual_data_type as same,
then it upto you if you want to implement read and write , as it is unneccesary to do it as the computer can hanle
it , by all means it can be used as preprocessing function.