Refined I_BINARYtrNode and added main.cpp
In binaryTree.h
================

Refined the interfaces

In I_BINARYtrNode, removed the data part , as the Interface only need
the function need for it's successor.

And changed it's template to hold only the data presented by the node's
key, rather than having both.

Added the height as to be the a pure virtual.

And most of the changes are due to change of template parameter in
I_BINARYtrNode.

In main.cpp
===========

Added as driver source file.

In NamingConvention.md
======================

Updated the # Pointer section to specify for pointer to node.