updated binaryTree.h
Also differentiated the Binary Search Tree Node into it's
base class and concrete class (is inheriated from the base class).

Created a design.md to specifiy the desgin logic used if it is
confused and if can improve upon that please do.

Updated NamingConvention.md to have Base class idea and it
representation in the library