AVL tree completed testing pending
In this AVL tree, it is unoptimized as I directly stored the height,
the efficient one would have been storing the height difference (not
extacly diffrence). In this optimized version we take 2 bit unsigned
int h, this will not directly store height rather relatively height
with it's sibling.

So, 2 bit will have (0,1,2,3), we will take equilibrum at 1. In other
words if sibling height is same the present h = 1, else h can vary
from 0 to 3. so, 1,2 are balancing conditions and 3,0 are unbalancing
one.

  0  - 1 -  2  -  3
 h-1   h   h+1   h+2

 where h is the common height, using this indication we can do. As the
 present is inefficient as the size of h increase as number of node
 increases.